requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
openai>=1.0.0
anthropic>=0.18.0
tiktoken>=0.5.0
numpy>=1.24.0
scikit-learn>=1.3.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
trafilatura>=2.0.0
pypdf>=6.0.0
rdflib>=7.0.0
SPARQLWrapper>=2.0.0
networkx>=3.0.0
gradio>=6.0.0
plotly>=5.0.0
markdown>=3.4.0
youtube-transcript-api>=0.6.0
html2text>=2024.0.0
python-docx>=1.0.0
markitdown[docx]>=0.1.4
sentence-transformers>=2.2.0
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.21.0
